stream-json: split modules, followed lint suggestions.
