+++ types/stream-json/streamers/StreamBase.d.ts
@@ -1,0 +1,17 @@
+import { Transform, TransformOptions } from 'stream';
+import * as Assembler from '../Assembler';
+
+export = StreamBase;
+
+declare class StreamBase extends Transform {
+    constructor(options?: StreamBase.StreamOptions);
+}
+
+declare namespace StreamBase {
+    type ObjectFilterFunction = (asm: Assembler) => boolean | undefined;
+
+    interface StreamOptions extends TransformOptions {
+        objectFilter?: ObjectFilterFunction;
+        includeUndecided?: boolean;
+    }
+}
