+++ types/stream-json/utils/emit.d.ts
@@ -1,0 +1,7 @@
+import { Writable, Duplex, Transform } from 'stream';
+
+export = emit;
+
+declare function emit<Stream extends Writable | Duplex | Transform>(stream: Stream): Stream;
+
+declare namespace emit {}
